pyproject.toml
[project]
name = "monitoring"
version = "0.1.0"
description = "Portfolio factor-exposure monitoring with Carino-linked contributions"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pandas>=2.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "python-dateutil>=2.8",
]

[project.scripts]
monitor = "monitor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/monitor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]
